Move all fixtures to fixtures.py
