+++ opsdroid/conftest.py
@@ -1,47 +1,7 @@
 """Pytest config for all opsdroid tests."""
-import pytest
-
-import asyncio
-import contextlib
-import socket
-
-from opsdroid.testing import opsdroid, mock_api  # noqa
-from opsdroid.connector import Connector
+# All global fixtures should be defined in the fixtures file
+from opsdroid.testing.fixtures import *  # noqa
 
 from opsdroid.cli.start import configure_lang
 
-__all__ = ["opsdroid"]
-
 configure_lang({})
-
-
-@pytest.fixture(scope="session")
-def get_connector():
-    def _get_connector(config={}):
-        return Connector(config, opsdroid=opsdroid)
-
-    return _get_connector
-
-
-@pytest.yield_fixture
-def event_loop():
-    """Create an instance of the default event loop for each test case."""
-    loop = asyncio.get_event_loop_policy().new_event_loop()
-    yield loop
-    loop.close()
-
-
-@pytest.fixture
-def bound_address(request):
-    s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-    with contextlib.suppress(socket.error):
-        if hasattr(socket, "SO_EXCLUSIVEADDRUSE"):  # only on windows
-            s.setsockopt(socket.SOL_SOCKET, socket.SO_EXCLUSIVEADDRUSE, 1)
-        if hasattr(socket, "SO_REUSEPORT"):  # not on windows
-            s.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEPORT, 0)
-        s.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 0)
-
-    host = request.param if hasattr(request, "param") else "0.0.0.0"
-    s.bind((host, 0))  # an ephemeral port
-    yield s.getsockname()
-    s.close()
